+++ src/main/java/com/inventory/detail/exception/InventoryNotFoundException.java
@@ -7,7 +7,7 @@ public class InventoryNotFoundException extends RuntimeException {
 	 */
 	private static final long serialVersionUID = 1L;
 
-	public InventoryNotFoundException(int id, String inventoryType){
+	public InventoryNotFoundException(Long id, String inventoryType){
 		super("NO INVENTORY FOUND OF TYPE "+inventoryType+ " FOR ID: "+ id);
 	}
 }
