+++ src/main/java/com/inventory/detail/service/InventoryService.java
@@ -9,12 +9,12 @@ public interface InventoryService {
 
 	public List<Inventory> getSelectedInventoryTypeDetails(String type) throws Exception;
 
-	public Inventory getSelectedItemDetail(String type, int id) throws InventoryNotFoundException, Exception;
+	public Inventory getSelectedItemDetail(String type, Long id) throws InventoryNotFoundException, Exception;
 
 	public void saveItemDetail(Inventory inventory) throws Exception;
 
-	public void updateItemDetail(Inventory inventory, String type, int id) throws InventoryNotFoundException, Exception;
+	public void updateItemDetail(Inventory inventory, String type, Long id) throws InventoryNotFoundException, Exception;
 
-	public void deleteItemDetail(String type, int id) throws InventoryNotFoundException, Exception;
+	public void deleteItemDetail(String type, Long id) throws InventoryNotFoundException, Exception;
 
 }
